Decent results
* reduced penalty for drift to penalize less drift than change size
* added penalty for area change
* improved main tracker reasoning for calling its different components

These changes were made so that the online classifier tracker (presented in Gu et al) performs better in environments with sparse features.
Indeed, when the frames don't have many features in the background, the tracker is likely to increase it size drastically to add nearby features that are wrongfully attributed to the tracked object.